feat: move to better algorithm for finding all instances of an LHS in the frontier
This could be in the worst case O(mn), where m is the length of the
frontier and n is the length of the lhs, but most of the time it should
be much better than this, plus n will never be too big, and this covers
all cases since it does a complete comparison. I could try backtracking
but my guess is the average time complexity would be roughly the same
then, and be a whole lot more complicated to implement and understand.